requirements.txt
torch
torchvision
diffusers
transformers
accelerate
peft
omegaconf
safetensors
Pillow
opencv-python
numpy
tqdm
wandb
ipython
